shorten the timeout between hiding and removal
When hiding overlay we keep it in the DOM tree for a short amount of
time to make sure that CSS transition is visible. For most browsers it
does not matter how long is that period, because our CSS prevents hidden
overlay from capturing any pointer events.

Since IE < 11 does not support `pointer-events` property for CSS[1],
hiding overlay keeps browser unresponsive for ~2s.

This commit shortens that period to 350ms, which should allow transition
to complete, and make IE usable a bit faster.

fixes #7
see #13 for a fix that removes hardcoded timeout

[1]: http://caniuse.com/#search=pointer-events